+++ CHANGES.md
@@ -11,3 +11,6 @@
 * rebrand to Git Remote Updater
 * keep retrying if remote response is WP_Error
 * added filter and function to remove slugs from appearing for updates
+* added a Settings tab to add site data, removed JSON storage directory
+* added nonce checks for WP_List_Table
+* move Git Remote Updater a top level menu